Backend/requirements.txt
requests
fastapi==0.115.*
uvicorn[standard]==0.30.*
pydantic==2.*
slowapi==0.1.9
httpx==0.27.*
python-dotenv==1.0.*
